Was changed package due to moving to deque package all related stuff
