+++ lab1/src/main/java/org/denys/hudymov/App.java
@@ -1,8 +1,10 @@
 package org.denys.hudymov;
 
+import org.denys.hudymov.deque.ConsoleInterface;
+
 public class App {
 
-    public static void main(String[] args){
+    public static void main(String[] args) {
         ConsoleInterface consoleInterface = new ConsoleInterface();
         consoleInterface.run();
     }
